requirements.txt
# AFSUAM Measurement System Requirements

# Core dependencies
numpy>=1.20.0
pandas>=1.3.0
scipy>=1.7.0

# Serial communication
pyserial>=3.5

# RFID Reader (LLRP)
sllurp>=0.5.0
twisted>=21.0.0

# Optional: Plotting
matplotlib>=3.4.0

# Optional: Machine Learning
scikit-learn>=0.24.0
joblib>=1.0.0
